refactor(frontend): separating the logics of  the form sales component
